refactor: Split parser/create.rs into modular structure
- Created 3 logical modules by parsing responsibility:
  * table.rs - Main CREATE TABLE parsing logic (98 lines)
  * types.rs - Data type parsing (189 lines)
  * constraints.rs - Constraint parsing (221 lines)
- Converted create.rs to mod.rs with re-exports (15 lines)
- All 410 parser tests passing
- No functional changes, pure refactoring

Reduces parser/create.rs from 567 lines to ~15 lines mod.rs + 3 focused modules
