+++ crates/parser/src/parser/create/mod.rs
@@ -1,0 +1,14 @@
+//! CREATE TABLE parsing
+//!
+//! This module handles parsing of CREATE TABLE statements, organized into:
+//! - `table` - Main CREATE TABLE parsing logic
+//! - `types` - Data type parsing (INTEGER, VARCHAR, NUMERIC, etc.)
+//! - `constraints` - Constraint parsing (PRIMARY KEY, FOREIGN KEY, etc.)
+
+mod constraints;
+mod table;
+mod types;
+
+pub use table::*;
+pub use types::*;
+pub use constraints::*;
